Add macOS package override for fsevent-sys
This commit should allow https://github.com/rust-analyzer/rust-analyzer
to build under macOS with `cargo2nix`. Previously, I was having some
issues compiling `rust-analyzer` on my Macbook Pro with
`cargo xtask install --server` on macOS Mojave because of a missing
symbol in the `CoreFoundation` framework:

```
  = note: Undefined symbols for architecture x86_64:
            "_CFURLResourceIsReachable", referenced from:
                fsevent_sys::core_foundation::str_path_to_cfstring_ref::he5fcd7d2a6a1a49d in libfsevent_sys-d6e1680613fbf498.rlib(fsevent_sys-d6e1680613fbf498.fsevent_sys.3i346u1f-cgu.4.rcgu.o)
          ld: symbol(s) not found for architecture x86_64
          clang-7: error: linker command failed with exit code 1 (use -v to see invocation)
```

Seems that this symbol exists upstream in the latest binary version of
the framework, but isn’t available in the Nixpkgs version of the Apple
SDK because Apple refuses to release the latest source ([reference]).

[reference]: https://discourse.nixos.org/t/help-with-rust-linker-error-on-darwin-cfurlresourceisreachable/2657/3

It turns out that recent versions of Nixpkgs contain a hack that pulls
in an impure version of `CoreFoundation` from the underlying system, but
it seems to only link as expected with `nix-build` and not `nix-shell`
plus Cargo.

This `CoreServices` dependency is coming from the `fsevent-sys` crate,
so adding this package override allows it to be built under `cargo2nix`,
which in turn, allows `rust-analyzer` to be built with `cargo2nix`.