Remove 'airplane mode' section in README.md for now
Since this feature is not on the roadmap to be reintroduced anytime
soon, we should remove this section from the README.md to avoid
confusing new users.

Originally, this feature was built as a more reliable alternative to
`cargo build --offline`. However, it was rather expensive to maintain
and prevented us from refactoring and simplifying the overlay UX leading
up to release 0.4.0, so it was removed. Additionally, `cargo build
--offline` has been working much better under more recent versions of
Rust, meaning there is less of a need to keep the Nix-based airplane
mode feature around.

If there is enough demand to reintroduce this feature, we can always
reconsider this decision post-1.0.