Document reporting crate build failures in CONTRIBUTING.md
This commit expands the CONTRIBUTING.md file to include a description of
how we handle bug reports of build failures of public crates available
on Crates.io, including where overrides can be found.

There are also some miscellaneous improvements to the verbiage to make
the text a bit less formal.