Convert Cargo.nix templating to Tera
This should greatly simplify our Rust rendering logic, with little
performance impact on release builds but a noticeable slowdown in debug
builds.

To test the integrity of the changes in this commit, a regression test
was created by wrapping `diesel`, a complex and feature-heavy crate
workspace, with `cargo2nix` and generating a `Cargo.nix` with the
latest `master` (0.8.1-unreleased) and a `Cargo.nix2` generated with
this commit.

Running `diff Cargo.nix Cargo.nix2` confirms that the only changes
between the two versions are:

1. The addition of a newline between the top comment and the Nix
   expression.
2. The removal of an unnecessary trailing space from each line of
   `rootFeatures`.
3. The unquoting of `unknown` in members of `workspace`.
4. The removal of an unnecessary trailing newline from the bottom of the
   file.

This means that this refactor can be considered a non-breaking change,
worthy of a patch version bump.