Use failure as application-wide error type
This matches the upstream convention used by Cargo with its
`CargoResult` type aliasing `failure::Error`. In later versions of
Cargo, `failure` has been replaced by `anyhow` (rust-lang/cargo#7776),
but this should suffice for the time being while we still depend on this
Cargo version.

The pattern of `try_main()` seen here is a common pattern seen in other
CLI tools such as `ripgrep`. Having `main()` return a `Result` isn't
useful here because it only shows the `Debug` output and not `Display`,
which typically results in a very poor experince for the end-user.