Specify crossSystem ? null for cargo2nix Rust crate
Since the `cargo2nix` crate cannot even compile under `musl` on Linux
anyway, thanks to the `nix-prefetch-git` dependency added in #62, it no
longer makes sense to recommend cross-compiling to that platform by
default. Instead, we should default to compiling for the system and
`libc` of the build platform, where the command is likely to be running
anyway.

This is not a breaking change for downstream users of the `cargo2nix`
overlay. It also should not affect future installations of the
`cargo2nix` CLI tool because the README already recommended installing
with `--arg crossSystem null` anyway.