Restructure second example project
As with the hello-world project, we take a more hands-on and less
intimidating approach to teaching `cargo2nix` by building a simple HTTP
client with `reqwest` and gently introducing the concept of overrides.

The next example will show how to Nix-ify a small crate workspace and
build and run some simple tests.