Make packageOverrides default to 'overrides.all'
As discussed in [this GitHub review thread], it's relatively rare for a
user to want to build a project with absolutely no overrides, so we
change the default setting for `packageOverrides` to `overrides.all`.

[this GitHub review thread]: https://github.com/tenx-tech/cargo2nix/pull/93#pullrequestreview-375921456

Nix is lazy, so it's not like it will attempt to apply any of the
overrides unless it actually needs to. It also doesn't increase the
fetch size of the `cargo2nix` overlay since it's already being fetched
in its entirety by the user.

If the user wants to build a project with absolutely no overrides, which
a select few might, they can always specify:

```nix
packageOverrides = _: [];
```