Remove empty link-flags files
Background is that when certain programs (home manager) are building a profile
to activate, these ./link-flags files are going to collide even though they
don't end up in the user's profile.  While this may be a buggy home manager
behavior, it would also be better not to have unnecessary outputs.

A proper fix needs to start by asking what these files are used for after
building the current crate and its dependents.  This file does not appear to
serve any standard role outside of the cargo2nix context and some of the
contents may be extraneous.

The derivation for cargo2nix will output a .link-flags with flags inside in an
otherwise empty lib folder while rust-analyzer with a busy lib folder has an
empty .link-flags

With the current home manager behavior, any two cargo2nix projects that leave
behind these identically named files will collide.  Even if empty ./link-flags
and ./link-flags with no rlib associations are removed, two cargo2nix
derivations that link to native libraries and have rlib outputs are enough to
trigger the collision.