Escape embedded TOML in profilesByName properly
The Tera templating code for `profilesByName` hadn't been rigorously
tested, and as such, it didn't perform proper escaping of sensitive
characters (e.g. `${`, `"`, and so on) before embedding the unsanitized
TOML into a double-quoted Nix string in the `Cargo.nix`. This string
might also contain newlines or other control characters, which would
need to be escaped as well.

This commit utilizes `String::escape_debug()` to perform the bulk of the
escaping work, sanitizing common control characters such as `\n`, `\r`,
and `\t`, and then we also replace all instances of `${` with `\${` to
avoid accidental string interpolation of variables.