Build examples against HEAD version of overlay
This ensures that we test our changes against the latest version of the
overlay and not the version pinned in the `examples`. This allows us to
test the proposed changes in PRs against more Cargo projects than
`cargo2nix` itself.

From this commit onward, the example projects are built with the
`--message-format=json` install method (Rust >=1.41) and `cargo2nix`
itself is built with the older install method (Rust <1.41) using the
runtime selection logic introduced a few commits ago.

This lets us test both the old and new code paths for regressions in CI.