Eliminate cp-related error spam from logs
There are two classes of `cp` failures being spammed in the logs, and
while they are both innocuous, they are quite annoying: (1) attempting
to copy `bin` or `lib` artifacts to `$out/{bin,lib}` when the crate
doesn't actually have such target(s) to speak of, and (2) when
attempting to copy artifacts pointed to by `out_dir` when the variable
is not actually set.

This commit fixes both of these by (1) redirecting the innocuous error
output to `/dev/null` and (2) gating a chunk of the copy logic on
whether the `out_dir` local was set successfully by `jq`.

As an aside, we also `rmdir "$out/bin"` if the directory is left empty.
By contrast, it appears that `$out/lib` is always populated by a
`.link-flags` file at minimum.