Udev override
Build.rs attmepts to directly invoke rustc as part of the build check instead of
using cargo.  Rustc is not on the path at this point.

It possibly makes sense to further cover this creative build.rs case by adding
rustc to the path in `setBuildEnv` hook directly and simplify this override to
just the native dep