Variable Sources
This patch makes it possible to separate generation of the Cargo.nix from the
distribution of the source that the Cargo.nix provides dependencies for.  This
is a necessary step in increasing the utility of Cargo2nix for generic packaging
of Rust software for nixpkgs.

The tera template is paramaterized to handle the edge case where the path
provided for concatenation with nix points to the source directory root, a
situation that leads to an erroneous expression of `$workspaceSrc + "."`;

In cases where the source is external and the local path at generation time is,
for instance, `crates/toolchain/`, the path concatenation will evaluate the
expression, `src = fetchCrateLocal workspaceSrc + "/crates/toolchain/"`, a valid
concatenation of a path literal.