Hack around filtering git source
Assuming things that don't belong in git sources are not being pulled (like
target etc), this hack is fairly innocuous.  However, it completely negates the
stated purpose and current behavior of `localPatterns` when the source is not a
path (because fetchFromGithHub etc are "set" types).  The example itself is
contradictory to the behavior introduced in this commit.  At the very least we
need a tracking issue if this is landed.