Fix Cargo link issue on mac
This fix provides libgit2 to the libgit2-sys crate that was experiencing a
linking issue with cargo later.  Same pattern of providing native deps so that
sys crates don't fail to build them with their limited knowledge of nixpkgs
excellence

Cargo 0.49 has no breaking changes and was also tested and works.  Not bumping
for now.  0.48 can read the new lockfile format just fine.