Flakify Hello-world Example
Went for shorter writing.  The existing guides have a tendency to say, "in the
next step will do X" and then follow up with "Now we will do X in this step".
Tried to make it more linear and the kind of writing people expect when diving
into a project and just wanting to see some success as soon as possible before
deciding to go deeper or move on.