De-duplicate link flags
These flags are combined from dependencies, resulting in link flag
amplification.  This change reduced rust analyzer's .link-flags from 191k to
2k.

It's a bit tricky because there are a lot of places we touch .link-flags.  The
change first writes the jq with newlines so that these files can be filtered
with sort -u.  When we read the flags we sort them.  When we write them, we sort
them.

There probably is still more I/O than there needs to be.  In dev, I found that
running the hook to set the environment took about 2-3 seconds.