Run examples with --update-input cargo2nix
Over time, this slightly alleviates the issue of not having subflakes

As the newer overlays hit master, CI will at least stay kind of up to date.
Updating the lock files is pretty tedious.  Maybe should just script it.

The only way to get out of bootstrapping corners with CI is to put a working rev
of an overlay into the branch and then point the flakes at that rev, pushing
this last