Explicit target setting
This change enables overriding the target for rustc, making it build different
than whatever `crossPlatform` and `platform` would otherwise go for.

The rust toolchain may not have such a target, so this feature is incomplete
where the rust toolchain needs to be selected with one target and told to build
for another (because it's a target that toolchain can be configured to cross
for)

Builds for cross should build identically if the target is set explicitly to the
same value as the crossSystem config.

note about explicit target