move overlay & overlays to top-level of outputs
this allows downstream flake inputs to simply use `cargo2nix.overlay` etc

! Flake show complains if overlays is a list

Might need to upgrade my nix... This is contradictory.  When calling nixpkgs,
overlays must be a list.  When exposing overlays as an output, it must be an
attrset.  This is absurd becuase the overlays output cannot be used in nixpkgs
directly.

Will file a bug later if I can't resolve this in newer nix

The schema clearly states that it can be a list or a set