Feature Logic Overhaul
Context

Features were being declared without condition (required) on workspace crates
when the default feature would not have enabled them.  The result was that in
most cargo2nix builds, all workspace level features were almost always enabled
even of `rootFeatures` was left to the default in makePackageSet'

To correct this, it was determined that the existing implementation was
insufficient.  In order to know when a feature will activate another feature, we
must resolve the workspace with each top level feature.

Including logic about what each individual workspace crate enables has been
removed.  This is because 1) it makes build caching less effective 2)
communicating which crates are being built is difficult and if possible would
require some work on the overlay.

After these changes, the only builds possible to express in makePackageSet' are
those with all crates built but any combination of features.

https://github.com/cargo2nix/cargo2nix/issues/211 Has more analysis / discussion
/ design rambling