!SPIKE Feature logic overhaul
It's a long story.  Go copy paste the issue notes into this commit.

It's not done yet because _way_ too many feature resolutions are done.  There's
a resolution for every top level package, for every feature.  However, without
information about _why_ features are enabled, this is actually the way to do it.

What is lost in this change and NOT expected to come back is that optional
dependencies will now only be optional if they are optional on features, not on
workspace crates.