Attempting to fix Darwin (again)
Probably libgit2-sys isn't finding the available libgit anymore and builds it
and does it a bit differently than libgit2_curl or whichever it is.

libiconv is supplied elsewhere to all Darwin derivations

libgit2 possibly the root of what looks like a determinism issue.. let's give cc
and the vendored version (hope there is one) a shot.