Find crate using grep (like a hax0r)
Eliminate IFD from readFile used to verify crate source location before build

This was done to find the specific source for a crate, which narrows the file
tree of what's available to the build.  However, since we must fetch and inspect
the source to make this decision, it becomes IFD.

This means we will use whole sources rather than paths within git sources.
However, there may be a workspace.  This workspace will interfere with how the
cargo2nix overlay articulates cargo.  We also have to remove any .cargo
configuration.  Any workspace level information is one more thing to worry
about.

The crate is "found" using grep to look for the name key.  This is not entirely
accurate but given that we don't expect large numbers of such keys or situations
where there are duplicate crates with different versions in a source tree, we
treat these cases as degenerate (ie looking at something that is not intending
to be a crate).

The check could be fragile in some yet undiscovered edge case where a name key
for a non-git source is not caught by the regex as written.  These cases should
be easy to discover.