rustcLinkFlags & rustcBuildFlags
There was some incomplete support for rustcBuildFlags and rustcLinkFlags that
was added without any immediate motivating example to check the behavior.

Well, it wasn't fully implemented and would not have worked.  This change passes
through flags to the rustc invocations used via $RUSTC.

rustcflags was renamed to rustcLinkFlags and NIX_RUSTC_LINK_BUILD_FLAGS was
changed to NIX_RUSTC_BUILD_FLAGS.

Linking is... kind of an inaccurate way to put it.  Non-build invocations of
Rustc use this flag.  Probably going to change it before merging.