boostrap shell.   cargo2nix requires a rustc to run
The binary itself is now wrapped, but the development shell doesn't put rustc &
cargo on the path.

`nix shell` has kind of an unclear use right now... I'm not sure what's going on
in nix land, what they want us to do with it.

Adding a new bootstrap artifact was an opportunity to not pollute the `nix
develop` workspaceShell but still provide a nice one-liner.