Overhaul Workspace Shell & Overrides
A bit of history.  The workspace shell started off as a quick solution to get
overrides to express some of their side effects within the user's development
shell.  Because it didn't work closely with the actual mechanisms of mkShell,
the chosen mechanism, from overrides to dev shell, was use of propagated
inputs.  This and some cargo culting had resulted in overrides mimicking other
overrides, lots of use of propagatedNativeBuildInputs.

First, workspace shell's overhaul is given the actual package set instead of the
noBuild set that is used when running tests.  (the tests themselves may need
similar overhaul).  The new shell has one simple job:  Make everything required
to build every crate, except rust crates themselves, expressed in environment
and available from the store.  This has to be done for the entire rust crate
DAG.

Second, once the workspace shell could correctly express dependencies and shell
hooks etc without the abuse of propagated inputs, the overrides themselves could
be made much more sensible, much more consistent with the rest of Nix.

subtractLists works fine for derivation filtering

The issue that lead to using attrSets was using the wrong crate outputs to
filter on.  After that was fixed, it was no longer necessary to construct
attrSets by keys.  If complexity of a dependency set (n^2 because of lib.unique)
becomes too high, the attrSet based implementation should become favored again.
This won't really show up to users for n < 1000 where n is the immediate
dependencies of crates that are not crates themselves.

Apple frameworks must be propagated?  Missing a setupHook?