automatically update the lockfile and support --locked
There is a problem because cargo2nix is not using the cargo binary here, which
could result in the lockfile being different than the project's lock file.

We want the behavior that cargo implements privately, to see that the lockfile
is out of date and to ask if the user wants to update it.

Cargo2nix could offload the work via metadata and json parsing, which may make
more sense as a cargo plugin.

Cargo2nix will be limited in its capability to work seemlessly with newer or
older projects if they generate lockfiles that clash with the output of
cargo2nix.  The use of the cargo library further cements the potential for
disparity.