Use entire workspace, begin refactor of toml munging
The newer findCrate logic is accurate enough that we don't need to provide the
source path, although we could re-visit this if some sources don't have a
workspace in the source root.  Cargo still needs to be able to decide to use
such a source in any case, or we can't generate the expression.