Sanitize workspace & package tomls
Instead of the existing whitelist style jq statements, a blacklist is produced
to eliminate keys that cause issues in sandbox builds or will conflict with
configuration via nix expressions

The workspace toml is reduced so that it becomes a workspace for just a single
member, the crate we want to build.

The crate toml is reduced separately.

The build happens from the workspace level.