Examples: use GitHub cargo2nix URL, not relative path
The users are shown one thing and told to do another.

The use of parent directories in flake URLs is unreliable (that's why
`get-flake` exists) and should be avoided in examples.

Referring to the parent directory creates a circular dependency in `flake.lock`
files. `nix flake lock` keeps changing the sources every time it's run because
the `flake.lock` files in the examples affect the checksum of the parent flake,
which in turn goes to `flake.lock` in the examples.

With GitHub references, the locks would get out of date only when a new commit
is merged to the branch used in the URL.