Support Workspace Metadata (#387)
This applies the changes from #333 to the `main` branch.

On the topic of branches, the readme says that users should follow
`unstable` to get the latest changes. But for a while people have been
merging to `main` instead. Maybe because `unstable` has branch
protections enabled, or maybe because people are used to using `main`.

Should we fast-forward `unstable` to `main`, and stop using `main`, or
should we update the documentation to tell people to follow `main`
instead of `unstable`?

---------

Co-authored-by: Psionik K <73710933+psionic-k@users.noreply.github.com>